app-convertix/backend/requirements.txt
Flask==3.1.1
flask-cors==6.0.1
python-dotenv==1.1.0
requests==2.32.3
SQLAlchemy==2.0.41
aiohttp==3.9.1